+++ app/src/main/java/com/erminesoft/my_account/myacount/db/DbManager.java
@@ -22,13 +22,13 @@ public final class DbManager implements DBbridge {
     }
 
     @Override
-     public void saveCostsToDb(int category, String nameText, double sumCost) {
+     public void saveCostsToDb(int category, String nameText, int sumCost) {
         ContentValues cv = Mapper.convertCosts(nameText, category, sumCost);
         baseHelper.getWritableDatabase().insert(DataBaseHelper.TABLE_COSTS, null, cv);
     }
 
     @Override
-    public void saveIncomeToDb(int category, String nameText, double sumIncome) {
+    public void saveIncomeToDb(int category, String nameText, int sumIncome) {
         ContentValues cv = Mapper.convertIncome(nameText, category, sumIncome);
         baseHelper.getWritableDatabase().insert(DataBaseHelper.TABLE_INCOME, null, cv);
     }
